test(opencode): consolidate session prompt tests into Effect style (#23710)
